requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.0.2
future==0.18.1
graphviz==0.19.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
libretranslatepy==2.1.1
line-bot-sdk==1.14.0
lxml==4.7.1
MarkupSafe==1.1.1
pygraphviz==1.5
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
transitions==0.6.9
translate==3.6.1
urllib3==1.25.6
Werkzeug==0.16.0
